Moved compiler into its own file, handled print warnings, and removed unused import from sdk.dart
