+++ ide/app/lib/services/compiler_service.dart
@@ -1,0 +1,64 @@
+// Copyright (c) 2014, Google Inc. Please see the AUTHORS file for details.
+// All rights reserved. Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+library spark.services_impl.compiler;
+
+import "dart:async";
+
+
+import '../dart/sdk.dart';
+import "../../services_impl.dart";
+import 'compiler.dart';
+import '../utils.dart';
+
+class CompilerServiceImpl extends ServiceImpl {
+  String get serviceId => "compiler";
+
+  DartSdk sdk;
+  Compiler compiler;
+
+  Completer<ServiceActionEvent> _readyCompleter =
+      new Completer<ServiceActionEvent>();
+
+  Future<ServiceActionEvent> get onceReady => _readyCompleter.future;
+
+  CompilerServiceImpl(ServicesIsolate isolate) : super(isolate);
+
+  Future<ServiceActionEvent> handleEvent(ServiceActionEvent event) {
+    switch (event.actionId) {
+      case "start":
+        // TODO(ericarnold): Start should happen automatically on use.
+        return _start().then((_) => new Future.value(event.createReponse(null)));
+        break;
+      case "dispose":
+        return new Future.value(event.createReponse(null));
+        break;
+      case "compileString":
+        return compiler.compileString(event.data['string'])
+            .then((CompilerResult result)  {
+              return new Future.value(event.createReponse(result.toMap()));
+            });
+        break;
+      default:
+        throw "Unknown action '${event.actionId}' sent to $serviceId service.";
+    }
+  }
+
+  Future _start() {
+    _isolate.chromeService.getAppContents('sdk/dart-sdk.bin').then((List<int> sdkContents) {
+      sdk = DartSdk.createSdkFromContents(sdkContents);
+      compiler = Compiler.createCompilerFrom(sdk);
+      _readyCompleter.complete();
+    }).catchError((error){
+      // TODO(ericarnold): Return error which service will throw
+      print("Chrome service error: $error ${error.stackTrace}");
+    });
+
+    return _readyCompleter.future;
+  }
+}
+
+
+// Used to avoid 'print hidden' warning
+void print(var message) => sendPrint(message);
